Refactor calendar components to utilize updated `useControlledValue` hook
- Updated imports of `useControlledValue` in `DateCalendar` and `DateRangeCalendar` components to point to the new location in `../../utils/hooks`.
- Enhanced type handling for the `selected` state to accommodate `null` values, ensuring better compatibility with the updated hook.
- Adjusted the `selected` prop to default to `undefined` when `null`, improving the handling of unselected states.
- Modified the `Default` story for `DateCalendar` to include example arguments for `defaultSelected` and `selected`, enhancing story clarity.
