+++ packages/components/src/calendar/stories/date-calendar.story.tsx
@@ -112,7 +112,12 @@ export default meta;
 
 type Story = StoryObj< typeof DateCalendar >;
 
-export const Default: Story = {};
+export const Default: Story = {
+	args: {
+		defaultSelected: new Date( 1751472000000 ),
+		selected: new Date( 1751558400000 ),
+	},
+};
 
 export const DisabledDates: Story = {
 	args: {
